changed to double

+++ ns2timetable.m
@@ -2,7 +2,7 @@ function TT = ns2timetable(NS)
 
 t = linspace(0,NS.MetaTags.DataPointsSec,NS.MetaTags.DataPoints);
 t = seconds(t);
-X = NS.Data'; X = single(X);
+X = NS.Data'; X = double(X);
 lbl = {NS.ElectrodesInfo.Label};
 lbl = upper(lbl);
